gtest proper creation of planner
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@5146 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ motion_planning/sbpl/src/discrete_space_information/nav2d/environment_nav2D.cpp
@@ -886,7 +886,7 @@ void EnvironmentNAV2D::GetPredsofChangedEdges(vector<nav2dcell_t>* changedcellsV
 {
 	nav2dcell_t cell;
 
-	for(int i = 0; i < changedcellsV->size(); i++)
+	for(int i = 0; i < (int)changedcellsV->size(); i++)
 	{
 		cell = changedcellsV->at(i);
 		preds_of_changededgesIDV->push_back(GetStateFromCoord(cell.x,cell.y));
